@claude implement feat-3151 move-service-and-create-internal-dtos
